skills/oss-image-uploader/requirements.txt
oss2>=2.18.0
python-dotenv>=1.0.0
